backend/requirements.txt
asgiref==3.11.1
attrs==25.4.0
Django==6.0.2
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.17.3
pillow==12.1.1
psycopg2-binary==2.9.11
pyrsistent==0.20.0
python-decouple==3.8
PyYAML==6.0.3
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
